Add files via upload
Items, Computer Controlled Characters, and Place table, when choosing the type, the input is at the lower right side of the screen. The record or index number is at the bottom left side. Items and Computer Controlled Characters show the name at the bottom center. Places show part of the description in the lower center. This is for when you use cursor up and cursor down to change the index number, you can see name or description to help you find the record to edit.
Also added TAB to activate GO# or Goto Record (Index#). The Clear Key goes to the very first record.