Add files via upload
Updated large screen editor to use the screen to edit large text fields instead of trying to edit the strings and print them onscreen. When pressing enter, the Atari screen codes need to be converted to ATASCII before copied into strings. I might need to convert this to ML, so it does it faster.